+++ src/midiEditorCommands.cpp
@@ -740,7 +740,7 @@ void cmdMidiMoveCursor(Command* command) {
 		fakeFocus = FOCUS_NOTE;
 		s << " " << format(
 			translate_plural("{} note", "{} notes", count), count);
-		int mutedCount = count_if(notes.begin(), notes.end(), [](auto note) { return note.muted; });
+		int mutedCount = (int)count_if(notes.begin(), notes.end(), [](auto note) { return note.muted; });
 		if (mutedCount > 0) {
 			// Translators: used when reporting the number of muted notes in a chord.
 			// {} will be replaced by the number of muted notes. E.g. "3 muted"
@@ -889,7 +889,7 @@ class SortedMidiControlChangeIterator {
 		}
 		// Finally, return the last sorted CC at this new position.
 		this->sortCCsAtPos();
-		this->sortedIndexAtPos = this->sortedCCsAtPos.size() - 1;
+		this->sortedIndexAtPos = (int)this->sortedCCsAtPos.size() - 1;
 		return this->current();
 	}
 
@@ -1119,7 +1119,7 @@ void moveToChord(int direction, bool clearSelection=true, bool select=true) {
 			// Translators: used when reporting the number of notes in a chord.
 			// {} will be replaced by the number of notes. E.g. "3 notes"
 			s << format(translate("{} notes"), count);
-			int mutedCount = count_if(notes.begin(), notes.end(), [](auto note) { return note.muted; });
+			int mutedCount = (int)count_if(notes.begin(), notes.end(), [](auto note) { return note.muted; });
 			if (mutedCount > 0) {
 				// Translators: used when reporting the number of muted notes in a chord.
 				// {} will be replaced by the number of muted notes. E.g. "3 muted"
@@ -2306,8 +2306,8 @@ void postMidiToggleMute(int command) {
 	vector<MidiNote> selectedNotes = getSelectedNotes(take);
 	// Get selected CCs.
 	vector<MidiControlChange> selectedCCs = getSelectedCCs(take);
-	int noteCount = selectedNotes.size();
-	int CCCount = selectedCCs.size();
+	int noteCount = (int)selectedNotes.size();
+	int CCCount = (int)selectedCCs.size();
 	int eventCount = noteCount + CCCount;
 	if (eventCount == 0) {
 		return;
@@ -2323,8 +2323,8 @@ void postMidiToggleMute(int command) {
 			}
 			s << getMidiNoteName(take, selectedNotes[0].pitch, selectedNotes[0].channel);
 		} else {
-			int mutedCount = count_if(selectedNotes.begin(), selectedNotes.end(), [](auto note) { return note.muted; });
-			int unmutedCount = noteCount - mutedCount;
+			int mutedCount = (int)count_if(selectedNotes.begin(), selectedNotes.end(), [](auto note) { return note.muted; });
+			int unmutedCount = (int)noteCount - mutedCount;
 			if (mutedCount > 0) {
 				// Translators: used when reporting the number of muted notes.
 				// {} will be replaced by the number of muted notes. E.g. "3 notes muted"
@@ -2349,7 +2349,7 @@ void postMidiToggleMute(int command) {
 			}
 			s << describeCC(take, cc);
 		} else {
-			int mutedCount = count_if(selectedCCs.begin(), selectedCCs.end(), [](auto cc) { return cc.muted; });
+			int mutedCount = (int)count_if(selectedCCs.begin(), selectedCCs.end(), [](auto cc) { return cc.muted; });
 			int unmutedCount = CCCount - mutedCount;
 			if (mutedCount > 0) {
 				// Translators: used when reporting the number of muted CCs.
@@ -2365,8 +2365,8 @@ void postMidiToggleMute(int command) {
 			}
 		}
 	} else { // If both notes and CCs are selected
-		int mutedNoteCount = count_if(selectedNotes.begin(), selectedNotes.end(), [](auto note) { return note.muted; });
-		int mutedCCCount = count_if(selectedCCs.begin(), selectedCCs.end(), [](auto cc) { return cc.muted; });
+		int mutedNoteCount = (int)count_if(selectedNotes.begin(), selectedNotes.end(), [](auto note) { return note.muted; });
+		int mutedCCCount = (int)count_if(selectedCCs.begin(), selectedCCs.end(), [](auto cc) { return cc.muted; });
 		int mutedCount = mutedNoteCount + mutedCCCount;
 		int unmutedCount = eventCount - mutedCount;
 		if (mutedCount > 0) {
@@ -2433,7 +2433,7 @@ int midiStepTranslateAccel(MSG* msg, accelerator_register_t* accelReg) {
 	int oldCount;
 	MIDI_CountEvts(take, &oldCount, nullptr, nullptr);
 	// F1 is the note at the pitch cursor, f2 is 1 semitone above, etc.
-	const int relativeNote = msg->wParam - VK_F1;
+	const int relativeNote = (int)msg->wParam - VK_F1;
 	// If the shift key is being held, the cursor is not advancing, so we should
 	// not report the new position.
 	const bool reportNewPos = !(GetAsyncKeyState(VK_SHIFT) & 0x8000);
